% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almagro.R
\name{almagro}
\alias{almagro}
\title{The CODEX index in COPD}
\usage{
almagro(charlson_index, fev, mmrc, exacerbation)
}
\arguments{
\item{charlson_index}{The Charlson Index relates long-term mortality with the patient’s comorbidity. In general, no comorbidity is considered to be: 0-1 point, low comorbidity: 2 points, and high > 3 points.}

\item{fev}{A number for forced expiratory volume in one second (FEV1) expressed as percentage}

\item{mmrc}{A number for the dyspnea scale of modified Medical Research Council (mMRC)}

\item{exacerbation}{Severe exacerbations of COPD during the previous year (hospitalization or ED consultations).}
}
\value{
the CODEX index
}
\description{
The CODEX index in COPD
}
\examples{
almagro(charlson_index=1, fev=65, mmrc=1, exacerbation=2)
}
